student observation bug fixes

See merge request k12dev/erp/erp-revamp-frontend!283

+++ src/containers/assessment-central/create-question-paper-new/sections.js
@@ -155,7 +155,7 @@ const Sections = ({
   const getMarks = () => {
     let marks = 0;
     for (let i = 0; i < section?.mandatory_questions; i++) {
-      marks += parseInt(section?.test_marks[i]?.question_mark[0]);
+      marks += parseFloat(section?.test_marks[i]?.question_mark[0]);
     }
     return marks;
   };
